requirements.txt
# requirements.txt
streamlit
openai
firebase-admin
requests
pyrebase4
